# docs/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MiniMarket Plus API",
    "description": "Backend para gestion de inventario, ventas, carrito y usuarios con seguridad JWT.",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:9090",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/inventario/{id}": {
      "get": {
        "tags": [
          "Inventario"
        ],
        "summary": "Obtiene un movimiento de inventario por id con enlaces HATEOAS",
        "description": "Consulta un movimiento especifico de inventario y entrega enlaces hacia el recurso actual, la coleccion y el producto asociado.",
        "operationId": "obtenerMovimientoPorId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador unico del movimiento de inventario",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "403": {
            "description": "Usuario sin permisos suficientes"
          },
          "200": {
            "description": "Movimiento encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelInventarioDTO"
                }
              }
            }
          },
          "401": {
            "description": "Usuario no autenticado"
          },
          "500": {
            "description": "Error interno del servidor"
          },
          "404": {
            "description": "Movimiento no encontrado"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Inventario"
        ],
        "summary": "Actualiza un movimiento de inventario",
        "description": "Modifica los datos de un movimiento de inventario existente.",
        "operationId": "actualizarMovimiento",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador unico del movimiento de inventario",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Inventario"
              }
            }
          },
          "required": true
        },
        "responses": {
          "403": {
            "description": "Usuario sin permisos suficientes"
          },
          "401": {
            "description": "Usuario no autenticado"
          },
          "500": {
            "description": "Error interno del servidor"
          },
          "200": {
            "description": "Movimiento actualizado correctamente",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelInventarioDTO"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud invalida"
          },
          "404": {
            "description": "Movimiento no encontrado"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Inventario"
        ],
        "summary": "Elimina un movimiento de inventario",
        "description": "Elimina un movimiento registrado cuando existe en la base de datos.",
        "operationId": "eliminarMovimiento",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador unico del movimiento de inventario",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "403": {
            "description": "Usuario sin permisos suficientes"
          },
          "401": {
            "description": "Usuario no autenticado"
          },
          "500": {
            "description": "Error interno del servidor"
          },
          "204": {
            "description": "Movimiento eliminado correctamente"
          },
          "404": {
            "description": "Movimiento no encontrado"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/inventario": {
      "get": {
        "tags": [
          "Inventario"
        ],
        "summary": "Lista movimientos de inventario con enlaces HATEOAS",
        "description": "Retorna movimientos de inventario como CollectionModel\u003CEntityModel\u003CInventarioDTO\u003E\u003E, incluyendo links self, collection y producto.",
        "operationId": "listarMovimientosDeInventario",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Numero de pagina a consultar, comenzando desde 0",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            },
            "example": 0
          },
          {
            "name": "size",
            "in": "query",
            "description": "Cantidad de movimientos por pagina",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            },
            "example": 10
          }
        ],
        "responses": {
          "403": {
            "description": "Usuario sin permisos suficientes"
          },
          "401": {
            "description": "Usuario no autenticado"
          },
          "500": {
            "description": "Error interno del servidor"
          },
          "200": {
            "description": "Movimientos consultados correctamente",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionModelEntityModelInventarioDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Inventario"
        ],
        "summary": "Registra un movimiento de inventario",
        "description": "Crea una entrada o salida de inventario asociada a un producto.",
        "operationId": "registrarMovimiento",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Inventario"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Movimiento registrado correctamente",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelInventarioDTO"
                }
              }
            }
          },
          "403": {
            "description": "Usuario sin permisos suficientes"
          },
          "401": {
            "description": "Usuario no autenticado"
          },
          "500": {
            "description": "Error interno del servidor"
          },
          "400": {
            "description": "Solicitud invalida"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/ventas/{id}": {
      "get": {
        "tags": [
          "venta-controller"
        ],
        "operationId": "obtenerVentaPorId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Venta"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Rol": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string"
          }
        }
      },
      "Usuario": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "writeOnly": true
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Rol"
            },
            "uniqueItems": true
          }
        }
      },
...
      "Categoria": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string"
          }
        }
      },
      "Producto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string"
          },
          "precio": {
            "type": "number",
            "format": "double"
          },
          "stock": {
            "type": "integer",
            "format": "int32"
          },
          "categoria": {
            "$ref": "#/components/schemas/Categoria"
          }
        }
      },
...
      "Inventario": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "producto": {
            "$ref": "#/components/schemas/Producto"
          },
          "cantidad": {
            "type": "integer",
            "format": "int32"
          },
          "tipoMovimiento": {
            "type": "string"
          },
          "fechaMovimiento": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "EntityModelInventarioDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Identificador unico del movimiento de inventario",
            "example": 1
          },
          "productoId": {
            "type": "integer",
            "format": "int64",
            "description": "Identificador del producto afectado por el movimiento",
            "example": 1
          },
          "productoNombre": {
            "type": "string",
            "description": "Nombre del producto afectado por el movimiento",
            "example": "Arroz grano largo 1kg"
          },
          "cantidad": {
            "type": "integer",
            "format": "int32",
            "description": "Cantidad ingresada o descontada del inventario",
            "example": 50
          },
          "tipoMovimiento": {
            "type": "string",
            "description": "Tipo de movimiento realizado",
            "example": "Entrada"
          },
          "fechaMovimiento": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha y hora del movimiento de inventario",
            "example": "2026-07-18T19:00:00.000+00:00"
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        }
      },
...
      "Venta": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "usuario": {
            "$ref": "#/components/schemas/Usuario"
          },
          "fecha": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "CollectionModelEntityModelInventarioDTO": {
        "type": "object",
        "properties": {
          "_embedded": {
            "type": "object",
            "properties": {
              "inventarios": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/EntityModelInventarioDTO"
                }
              }
            }
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        }
      },
...
      "Link": {
        "type": "object",
        "properties": {
          "href": {
            "type": "string"
          },
          "hreflang": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "deprecation": {
            "type": "string"
          },
          "profile": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "templated": {
            "type": "boolean"
          }
        }
      },
      "Links": {
        "type": "object",
        "additionalProperties": {
          "$ref": "#/components/schemas/Link"
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}